+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-No changes yet.
+### Added
+- Structs now include `IsSet*` methods for fields that can be nil.
 
 ## [1.13.1] - 2018-10-04
 ### Fixed